+++ app/assets/stylesheets/exercises.scss
@@ -25,8 +25,10 @@
   flex-direction: column;
 }
 
-.no-exercises {
+.nothing-toshow {
   font-size: 40px;
+  margin-top: 250px;
+  text-align: center;
 }
 
 .g_exercises_main {
