+++ app/assets/stylesheets/groups.scss
@@ -13,12 +13,6 @@
   flex-direction: column;
 }
 
-.no-groups {
-  font-size: 40px;
-  position: relative;
-  top: -30px;
-}
-
 .container-scroll {
   position: relative;
   top: 10px;
