+++ app/assets/stylesheets/navbar.scss
@@ -157,7 +157,6 @@
   margin-left: 25px;
 }
 
-/******************/
 .dropdown:hover .dropdown-content {
   display: block;
 }
