+++ app/views/exercises/_empty.html.erb
@@ -1,4 +1,4 @@
-<div class="no-exercises my-auto mx-auto text-center">
+<div class="nothing-toshow mx-auto">
   <h4 class="">Nothing to show</h4>
   <i class="far fa-sad-tear"></i>
 </div>
