--- app/views/groups/_empty.html.erb
@@ -1,4 +1,0 @@
-<div class="no-groups my-auto mx-auto text-center">
-  <h4 class="">Nothing to show</h4>
-  <i class="far fa-sad-tear"></i>
-</div>
