+++ app/views/groups/show.html.erb
@@ -1,9 +1,9 @@
 <%= render 'layouts/navbar' %>
-<div class="myg_exercises-container">
-  <div class="d-flex flex-column justify-content-between" style="height: 100%;">
-    <div class="g_exercises_main"><%= check_gexercises_empty %></div>
-    <div class="exercise_add_new">
-      <%= link_to "ADD NEW EXERCISE", new_exercise_path, class: 'exercise_new_btn text-decoration-none text-light' %>
-    </div>
+<div class="myexercises-container">
+  <div class="container-scroll mb-2">
+    <%= check_gexercises_empty %>
+  </div>
+  <div class="exercise_add_new">
+    <%= link_to "ADD NEW EXERCISE", new_exercise_path, class: 'exercise_new_btn text-decoration-none text-light' %>
   </div>
 </div>
