+++ test/controllers/exercises_controller_test.rb
@@ -17,13 +17,10 @@ class ExercisesControllerTest < ActionDispatch::IntegrationTest
 
   test 'should create exercise' do
     assert_difference('Exercise.count') do
-      post exercises_url, params: { exercise: {
-                                                group_id: @exercise.group_id,
+      post exercises_url, params: { exercise: { group_id: @exercise.group_id,
                                                 name: @exercise.name,
                                                 time: @exercise.time,
-                                                user_id: @exercise.user_id 
-                                              }
-                                  }
+                                                user_id: @exercise.user_id } }
     end
 
     assert_redirected_to exercise_url(Exercise.last)
@@ -40,13 +37,10 @@ class ExercisesControllerTest < ActionDispatch::IntegrationTest
   end
 
   test 'should update exercise' do
-    patch exercise_url(@exercise), params: { exercise: {
-                                                          group_id: @exercise.group_id,
-                                                          name: @exercise.name,
-                                                          time: @exercise.time,
-                                                          user_id: @exercise.user_id
-                                                        }
-                                            }
+    patch exercise_url(@exercise), params: { exercise: { group_id: @exercise.group_id,
+                                                         name: @exercise.name,
+                                                         time: @exercise.time,
+                                                         user_id: @exercise.user_id } }
     assert_redirected_to exercise_url(@exercise)
   end
